requirements.txt
flask==2.3.3
docker==6.1.3
python-dotenv==1.0.0 